news/requirements.txt
amqp==1.4.9
anyjson==0.3.3
beautifulsoup4==4.5.1
billiard==3.3.0.23
celery==3.1.24
cssselect==0.9.2
Django==1.10.2
feedfinder2==0.0.4
feedparser==5.2.1
gunicorn==19.6.0
idna==2.1
jieba3k==0.35.1
kombu==3.0.37
lxml==3.6.4
newspaper3k==0.1.7
nltk==3.2.1
numpy==1.11.2
Pillow==3.4.1
python-dateutil==2.5.3
pytz==2016.7
PyYAML==3.12
redis==2.10.5
requests==2.11.1
requests-file==1.4
scikit-learn==0.18
scipy==0.18.1
six==1.10.0
sklearn==0.0
tldextract==2.0.1
